+++ .gitignore
@@ -2,6 +2,7 @@ data
 logs
 output
 src
+predictions
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
