+++ kilt/configs/retriever/default_blink.json
@@ -10,5 +10,7 @@
     "crossencoder_config": "/checkpoint/fabiopetroni/KILT/retriever/BLINK/crossencoder_wiki_large.json",
     "fast": false,
     "output_path": "logs/",
-    "top_k": 100
+    "top_k": 100,
+    "faiss_index": null,
+    "index_path": null
 }
